requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
agno
Pillow>=10.0.0
google-generativeai>=0.8.0
google-genai
python-docx>=1.1.0
python-dotenv>=1.0.0
pydicom>=2.4.0
numpy>=1.24.0,<2.0.0
